fix(shared+desktop): dev-time ESM resolution — channels 目录化 + 补漏 workspace bundle list
发起人 M1 收尾 dev 模式起不来,根因两处叠加:

## 根因

### A. channels.ts + channels/ 目录共存(本 #30C 重构引入)

#30C 拆 channels 时新建 `packages/shared/src/channels/` 8 个 domain 文件,但旧
`channels.ts` 文件没删 → vite resolver 看到 `import "./channels"` 时含糊
(同时存在文件 + 目录),触发 `Cannot find module '/.../packages/shared/src/channels'`。

### B. OPENTRAD_WORKSPACE_DEPS 漏 skill-runtime / risk-gate(M1 #23 / #28 引入)

`apps/desktop/electron.vite.config.ts` 的 `OPENTRAD_WORKSPACE_DEPS` 列表是 M0 #15
建立的(只 3 个 workspace deps)。M1 后期加 `@opentrad/skill-runtime`(#23 / #43
切换 PR)+ `@opentrad/risk-gate`(#28 阶段 2)时**没同步加进 list** → 它们被
externalize → main 进程运行时 Node ESM resolver 找 `packages/<pkg>/src/index.ts`
→ 不 fallback `.ts` 扩展 → ERR_MODULE_NOT_FOUND。

**CI 永远绿**(vitest / tsc 走自己的 resolver,不用 Node ESM 严格模式) — 同 W1
ABI / preload zod 同性质 bug,"CI 绿但 dev 死"。

## 治标修(本 commit)

1. **删 packages/shared/src/channels.ts** 文件,内容(主入口聚合段)搬到
   `packages/shared/src/channels/index.ts`(单一权威 source,无歧义)
2. **packages/shared/package.json exports."./channels"** 路径同步:
   `./src/channels.ts` → `./src/channels/index.ts`
3. **electron.vite.config.ts OPENTRAD_WORKSPACE_DEPS** 加 `@opentrad/skill-runtime`
   + `@opentrad/risk-gate`(注释标 M1 #23 / #28 漏加)+ 加教训注释提醒未来加
   workspace dep 必须同步加进 list

## 治本(独立 PR,不阻塞 #51 / #52 合)

retrospective followup #9 已记。**扩 electron-abi-smoke job**(或新加 dev-smoke
job)— 三平台 fresh checkout 后 ELECTRON_RUN_AS_NODE 起 main bundle 跑 5 秒,
所有 import 链路实际 resolve。同 PR A v2 (#36) sentinel 治本思路。

## 验证(治标 commit 已实测)

- `pnpm typecheck` 全 workspace 0 错(包括 desktop)
- `pnpm --filter @opentrad/desktop build` 三段全过(main 88 KB / preload 5.73 KB /
  renderer 2.49 MB)
- `pnpm dev` 起 12s 无 ERR_MODULE_NOT_FOUND,electron app 进程启动 OK
- preload import path `@opentrad/shared/channels` 不变(走 package exports → channels/index.ts)

## 不在本 commit

- 治本 dev-smoke job(独立 follow-up PR,见 retrospective followup #9)
- main 进程所有 ESM import 全面 audit(`grep -rn "from \"@opentrad/" apps/desktop/src/main/`)留 retrospective 段做

Refs #30
